Use GitHub actions for CI.

+++ test/MySql/AlterAuditTableCommand/testIntColumn/expected.sql
@@ -1,3 +1,3 @@
 alter table `test_audit`.`TABLE1`
-  change column `col1` `col1` int(8) null
+  change column `col1` `col1` bigint(20) null
 ;
