+++ test/MySql/AlterAuditTableCommand/testMultipleColumns/alter.sql
@@ -1,5 +1,5 @@
 alter table `TABLE1`
-  modify `col1`   int(8),
+  modify `col1`   bigint,
   modify `col10`  timestamp not null default now() on update now(),
   modify `col100` varchar(10) character set ascii collate ascii_bin
 ;
